support side effect function for swallow
Useful if want to pass in a logger to capture the error, even though for
business logic purposes you are OK to move along.
Should the effect function be wrapped in a try (EG below) to ensure it
doesn't throw (since we are in a swallow function after all), or is that
outside the intended scope of what is intended as swallowed?

We we did want to ensure the effect function is swallowed, to ensure for
sync and async code, we could just:

```
try {
  await effect();
} catch () {}
```